requirements.txt
numpy
tensorflow-gpu==1.9.0
opencv-python
requests
googledrivedownloader
youtube-dl
Pillow